chore(common): improve dimension preset related code
